+++ src/constants/constantSettings.ts
@@ -3,4 +3,4 @@ export const FIRST_OPEN = 'firstOpen.json'
 export const SCAM_WARNING = 'scamWarning.json'
 export const SETTINGS_PERMISSION_LIMITS = 'SETTINGS_PERMISSION_LIMIT'
 export const SETTINGS_PERMISSION_QUANTITY = 3
-export const STICKY_CONFIG = 'remoteConfigSticky.json'
+export const LOCAL_EXPERIMENT_CONFIG = 'remoteConfigSticky.json'