+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- added: Experiment config support between gui and login-ui
+- added: Experiment config for "Create Account" and "Login" button labels
 - changed: Always show gas warning when enabling tokens
 - changed: Replace text 'plugins' with 'providers' in Buy/Sell
 - changed: Tweak the boot background color on Android.